worldquant-alpha-finder/requirements.txt
httpx
rich
click
python-dotenv
